pyproject.toml
[project]
name = "rest-api-lab6"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-marshmallow>=1.3.0",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "marshmallow-sqlalchemy>=1.4.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
